+++ docs/index.md
@@ -67,6 +67,7 @@ As of now, VTODOGenerator supports the following fields:
 |---|---|---|---|
 | dtstamp | string|Time stamp for creation of the TODO. If not supplied, current time will be used. |  DTSTAMP|
 | summary | string|Summary of the task. | SUMMARY|
+| uid| string|UID for the task. If none is provided, a random id is autogenerated.| UID|
 | due | string|Due date and time of the task. | DUE|
 | completed| string|Date and time of completion of task  | COMPLETED|
 | completion| string|How much the task has progressed in percentage.| PERCENT-COMPLETE|
@@ -132,7 +133,7 @@ related to:[
     }
 ]
 ```
-### RRULE
+### rrule
 
 rrule defines the recurrence of the task, and it is valuable for generating repeating tasks.
 
@@ -217,7 +218,7 @@ var todoData = {
     UNTIL: 03/01/2024,
     },
     recurrence:[
-        {
+         "2023-12-31": {
             description: "Don't drink too much",
             recurrenceid: "20231231T000000"
         }
@@ -227,12 +228,12 @@ var todoData = {
 }
 ```
 
-On Dec 31, 2023, the task instance will show up with the description "Don't drink too much".
+On Dec 31, 2023, the task instance will show up with the description "Don't drink too much" (on most clients).
 
 
 ## Functions
 
-## generate(skipVCALENDAR)
+### generate(skipVCALENDAR)
 
 Generates the VTODO.
