+++ mkdocs.yml
@@ -1,1 +1,1 @@
-site_name: My Docs
+site_name: VTODOGenerator
