+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "vtodogenerator",
-  "version": "1.0.3",
+  "version": "1.1.1",
   "description": " A library to generate VTODO compatible with RFC5545 ",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "babel src -d dist"
   },
   "repository": {
     "type": "git",
@@ -23,5 +24,10 @@
   "homepage": "https://github.com/intri-in/vtodogenerator#readme",
   "dependencies": {
     "moment": "2.29.4"
+  },
+  "devDependencies": {
+    "@babel/cli": "7.21.5",
+    "@babel/core": "7.21.8",
+    "@babel/preset-env": "7.21.5"
   }
 }
