+++ src/index.js
@@ -1,0 +1,2 @@
+import VTodoGenerator from "./VTodoGenerator";
+export default VTodoGenerator
